updates to azure_rm_postgresqldatabase

+++ lib/ansible/modules/cloud/azure/azure_rm_postgresqldatabase.py
@@ -17,9 +17,9 @@ DOCUMENTATION = '''
 ---
 module: azure_rm_postgresqldatabase
 version_added: "2.5"
-short_description: Manage PostgreSQL Database instance
+short_description: Manage PostgreSQL Database instance.
 description:
-    - Create, update and delete instance of PostgreSQL Database
+    - Create, update and delete instance of PostgreSQL Database.
 
 options:
     resource_group:
@@ -43,7 +43,6 @@ options:
 
 extends_documentation_fragment:
     - azure
-    - azure_tags
 
 author:
     - "Zim Kalinowski (@zikalino)"
@@ -53,33 +52,27 @@ author:
 EXAMPLES = '''
   - name: Create (or update) PostgreSQL Database
     azure_rm_postgresqldatabase:
-      resource_group: resource_group_name
-      server_name: server_name
-      name: database_name
-      charset: charset
-      collation: collation
+      resource_group: TestGroup
+      server_name: testserver
+      name: db1
 '''
 
 RETURN = '''
-state:
-    description: Current state of PostgreSQL Database
+id:
+    description:
+        - Resource ID
     returned: always
-    type: complex
-    contains:
-        id:
-            description:
-                - Resource ID
-            returned: always
-            type: str
-            sample: id
-        name:
-            description:
-                - Resource name.
-            returned: always
-            type: str
-            sample: name
+    type: str
+    sample: /subscriptions/ffffffff-ffff-ffff-ffff-ffffffffffff/resourceGroups/TestGroup/providers/Microsoft.DBforPostgreSQL/servers/testserver/databases/db1
+name:
+    description:
+        - Resource name.
+    returned: always
+    type: str
+    sample: db1
 '''
 
+import time
 from ansible.module_utils.azure_rm_common import AzureRMModuleBase
 
 try:
@@ -114,16 +107,13 @@ class AzureRMDatabases(AzureRMModuleBase):
                 required=True
             ),
             charset=dict(
-                type='str',
-                required=False
+                type='str'
             ),
             collation=dict(
-                type='str',
-                required=False
+                type='str'
             ),
             state=dict(
                 type='str',
-                required=False,
                 default='present',
                 choices=['present', 'absent']
             )
@@ -134,36 +124,34 @@ class AzureRMDatabases(AzureRMModuleBase):
         self.name = None
         self.parameters = dict()
 
-        self.results = dict(changed=False, state=dict())
+        self.results = dict(changed=False)
         self.mgmt_client = None
         self.state = None
         self.to_do = Actions.NoAction
 
         super(AzureRMDatabases, self).__init__(derived_arg_spec=self.module_arg_spec,
                                                supports_check_mode=True,
-                                               supports_tags=True)
+                                               supports_tags=False)
 
     def exec_module(self, **kwargs):
         """Main module execution method"""
 
-        for key in list(self.module_arg_spec.keys()) + ['tags']:
+        for key in list(self.module_arg_spec.keys()):
             if hasattr(self, key):
                 setattr(self, key, kwargs[key])
-            elif key == "charset":
-                self.parameters["charset"] = kwargs[key]
-            elif key == "collation":
-                self.parameters["collation"] = kwargs[key]
+            elif kwargs[key] is not None:
+                if key == "charset":
+                    self.parameters["charset"] = kwargs[key]
+                elif key == "collation":
+                    self.parameters["collation"] = kwargs[key]
 
         old_response = None
-        results = dict()
+        response = None
 
         self.mgmt_client = self.get_mgmt_svc_client(PostgreSQLManagementClient,
                                                     base_url=self._cloud_environment.endpoints.resource_manager)
 
-        try:
-            resource_group = self.get_resource_group(self.resource_group)
-        except CloudError:
-            self.fail('resource group {0} not found'.format(self.resource_group))
+        resource_group = self.get_resource_group(self.resource_group)
 
         old_response = self.get_postgresqldatabase()
 
@@ -186,27 +174,36 @@ class AzureRMDatabases(AzureRMModuleBase):
             self.log("Need to Create / Update the PostgreSQL Database instance")
 
             if self.check_mode:
+                self.results['changed'] = True
                 return self.results
 
-            self.results['state'] = self.create_update_postgresqldatabase()
+            response = self.create_update_postgresqldatabase()
+
             if not old_response:
                 self.results['changed'] = True
             else:
-                self.results['changed'] = old_response.__ne__(self.results['state'])
-
-            # remove unnecessary fields from return state
-            self.results['state'].pop('type', None)
-            self.results['state'].pop('charset', None)
-            self.results['state'].pop('collation', None)
+                self.results['changed'] = old_response.__ne__(response)
             self.log("Creation / Update done")
         elif self.to_do == Actions.Delete:
             self.log("PostgreSQL Database instance deleted")
-            self.delete_postgresqldatabase()
             self.results['changed'] = True
+
+            if self.check_mode:
+                return self.results
+
+            self.delete_postgresqldatabase()
+            # make sure instance is actually deleted, for some Azure resources, instance is hanging around
+            # for some time after deletion -- this should be really fixed in Azure
+            while self.get_postgresqldatabase():
+                time.sleep(20)
         else:
             self.log("PostgreSQL Database instance unchanged")
-            self.results['state'] = old_response
             self.results['changed'] = False
+            response = old_response
+
+        if response:
+            self.results["id"] = response["id"]
+            self.results["name"] = response["name"]
 
         return self.results
 
@@ -219,10 +216,10 @@ class AzureRMDatabases(AzureRMModuleBase):
         self.log("Creating / Updating the PostgreSQL Database instance {0}".format(self.name))
 
         try:
-            response = self.mgmt_client.databases.create_or_update(self.resource_group,
-                                                                   self.server_name,
-                                                                   self.name,
-                                                                   self.parameters)
+            response = self.mgmt_client.databases.create_or_update(resource_group_name=self.resource_group,
+                                                                   server_name=self.server_name,
+                                                                   database_name=self.name,
+                                                                   parameters=self.parameters)
             if isinstance(response, AzureOperationPoller):
                 response = self.get_poller_result(response)
 
@@ -239,9 +236,9 @@ class AzureRMDatabases(AzureRMModuleBase):
         '''
         self.log("Deleting the PostgreSQL Database instance {0}".format(self.name))
         try:
-            response = self.mgmt_client.databases.delete(self.resource_group,
-                                                         self.server_name,
-                                                         self.name)
+            response = self.mgmt_client.databases.delete(resource_group_name=self.resource_group,
+                                                         server_name=self.server_name,
+                                                         database_name=self.name)
         except CloudError as e:
             self.log('Error attempting to delete the PostgreSQL Database instance.')
             self.fail("Error deleting the PostgreSQL Database instance: {0}".format(str(e)))
@@ -257,9 +254,9 @@ class AzureRMDatabases(AzureRMModuleBase):
         self.log("Checking if the PostgreSQL Database instance {0} is present".format(self.name))
         found = False
         try:
-            response = self.mgmt_client.databases.get(self.resource_group,
-                                                      self.server_name,
-                                                      self.name)
+            response = self.mgmt_client.databases.get(resource_group_name=self.resource_group,
+                                                      server_name=self.server_name,
+                                                      database_name=self.name)
             found = True
             self.log("Response : {0}".format(response))
             self.log("PostgreSQL Database instance : {0} found".format(response.name))
